refactor(spec-140): W2.5 — split validator test monolith (test-driven, D-140-07 gate)
Split `tests/unit/test_validator.py` (2,554 LOC, 127 tests) into 9 files
under `tests/unit/validator/` plus a shared `conftest.py`. Every test is
preserved — the post-split collection still reports 127 tests and they
pass at the same baseline as before (6 pre-existing migration-sha drift
failures unchanged). Categories are now addressable in isolation and the
heaviest single file (`test_mirror_sync_categories.py`) is 541 LOC.

Adds `tests/unit/validator/test_category_public_api.py` (W2.5.T6): an
AST-based parity guard that walks every `.py` file in the repo for
`from ai_engineering.validator.categories[...] import ...` statements
and asserts each imported symbol still resolves on its declaring
module. Two complementary tests confirm the seven category check
functions remain callable from the package and that
`tools/skill_app/lint_service` re-exports the canonical public-API set.

W2.5.T1 (split `manifest_coherence.py` into a per-dimension package)
and W2.5.T2 (split `mirror_sync.py` along the per-mirror seam) are
DEFERRED. Splitting these monoliths adds ~80-150 LOC of import /
`__init__.py` re-export scaffolding without test deletions to offset
the production growth at the D-140-07 2× ratio. Re-attempt only after
a clear ROI use case lands.

W2.5.T3 (delete `cross_references.py` 5-LOC stub) is BLOCKED. The spec
claim "no consumer" was incorrect: `tools/skill_app/lint_service.py:31`
imports `_check_cross_references` via the canonical
`ai_engineering.validator.categories` re-export surface, and
`tests/integration/test_gap_fillers4.py:388` exercises the same path.
Deleting the stub would break both consumers.

W2.5.T4 (absorb remaining 5-LOC stubs) is DEFERRED behind W2.5.T1 /
W2.5.T2.

LOC delta (W2.5 only):

| File                                       | Pre   | Post  | Delta |
| ------------------------------------------ | ----- | ----- | ----- |
| tests/unit/test_validator.py               | 2,554 | 0     | -2554 |
| tests/unit/validator/conftest.py           | 0     | 415   | +415  |
| tests/unit/validator/test_parse_counter_and_report.py | 0 | 206 | +206 |
| tests/unit/validator/test_file_existence_categories.py | 0 | 210 | +210 |
| tests/unit/validator/test_mirror_sync_categories.py | 0 | 541 | +541 |
| tests/unit/validator/test_counter_accuracy_categories.py | 0 | 305 | +305 |
| tests/unit/validator/test_manifest_coherence_categories.py | 0 | 309 | +309 |
| tests/unit/validator/test_skill_frontmatter_categories.py | 0 | 212 | +212 |
| tests/unit/validator/test_validate_content_integrity.py | 0 | 48 | +48 |
| tests/unit/validator/test_shared_utilities.py | 0 | 309 | +309 |
| tests/unit/validator/test_category_public_api.py | 0 | 155 | +155 |
| **Net test delta**                         |       |       | **+156** |
| **Net production delta**                   |       |       | **0**    |

D-140-07 gate (test deletion >= 2 x production-LOC overhead): production
overhead is 0; the gate is vacuously satisfied. The +156 LOC test growth
is the necessary boilerplate for splitting a 2,554-LOC monolith into
10 files (per-file module docstrings, imports, test-class skeletons)
plus the 155-LOC parity test that is itself a new W2.5 deliverable.
Excluding the parity test, the test-split overhead is +1 LOC.

Ai-Eng-Gate: passed
